+++ packages/isolate-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ricsam/isolate-server
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @ricsam/isolate-client@0.1.21
+
 ## 0.2.3
 
 ### Patch Changes